\name{cooccurrences}
\alias{cooccurrences}

\title{ compute association score beteen tokens }
\description{
Compute association between pairs of words in a text. Not implemented yet.
}
\usage{
cooccurrences(corpus)
}

\arguments{
  \item{corpus}{ The number of cooccurrences found in a given unit (phrase, sentence, paragraph) between \code{type1} et \code{type2} }
}


\value{
  
}

\references{ 
Lafon P. (1981) <<~Analyse lexicom\'etrique et recherche des cooccurrences~>>, \emph{Mots}, pp. 95--148, n~3.
See http://www.persee.fr/showPage.do?urn=mots\_0243-6450\_1981\_num\_3\_1\_1041
}

\author{ Sylvain Loiseau }

\examples{

}
