\name{corpus.tiny.cooccurrences}
\alias{corpus.tiny.cooccurrences}
\docType{data}
\title{ testing data set for cooccurrences function }
\description{

  A vary basic representation of a text for illustrating cooccurrences computing

}
\usage{data(corpus.tiny.cooccurrences)}
\format{
}
\details{

  Data frame containing three values, "S" for sentences boundaries, "G" and
  "F" for two phenomena whose cooccurrences inside sentences are observed.

}
\source{

  Lafon P. (1981) Lexicom \' e trie et cooccurrences, Mots, 2, pp. 120.

}
\references{

  Lafon P. (1981) Lexicom \' e trie et cooccurrences, Mots, 2, pp. 95-148.

}
\examples{

data(corpus.tiny.cooccurrences)
#cooccurrences.corpus(corpus.tiny.cooccurrences);

}
\keyword{datasets}
