\name{specificites.lexicon.probabilities}
\Rdversion{1.1}
\alias{specificites.lexicon.probabilities}

\title{
  Raw hypergeometric probabilities
}
\description{
  Raw hypergeometric probabilities for computing specificites, see \code{\link{specificites.lexicon}}
}
\usage{
  specificites.lexicon.probabilities(lexicon, sublexicon)
}

\arguments{
  \item{lexicon}{See \code{\link{specificites.lexicon}}}
  \item{sublexicon}{See \code{\link{specificites.lexicon}}}
}

\value{
  Hypergeometric probabilities. See \code{\link{specificites.lexicon}}.
}

\author{ Sylvain Loiseau }

\seealso{
  See \code{\link{specificites.lexicon}}
}
