\name{specificites.probabilities}
\Rdversion{1.1}
\alias{specificites.probabilities}

\title{
  Raw hypergeometric probabilities
}
\description{
  Raw hypergeometric probabilities, used for computing specificites (see \code{\link{specificites}}).
}

\usage{
specificites.probabilities(lexicaltable, types = NULL, parts = NULL)
}

\arguments{

  \item{lexicaltable}{ see \code{\link{specificites}} }

  \item{types}{ see \code{\link{specificites}} }

  \item{parts}{ see \code{\link{specificites}} }

}

\seealso{

 see \code{\link{specificites}}.

}
